feat: Add user status toggling functionality (feat/admin/interface)
Added functionality to toggle user status in the admin
interface.  This includes a new toggle button in the
user list and updates to the AdminController to handle
requests.  The UserRepository is now injected into the
AdminController.  The user.html template now displays
user email addresses and the new toggle button.